Extract Blockchain intot its own module
